requirements.txt
py-cord>=2.4.0
python-dotenv>=0.19.2
aiohttp>=3.8.1
fastapi>=0.95.0
uvicorn>=0.17.6
pydantic>=1.9.0
APScheduler>=3.9.1